I'm back
Well this is a suprise. I'm back. I've changed the name of the bot and made a new VPS.
The main reason is because it is now winter and I'm going to be spending more time in my house so
I need to start having things to do so I don't start going mentally insane. I'm probably going to
change up this bot completly but for the time being I'm just going to run the old code. I might start
rewriting this in Java. This will likely be followed by other commits as some of the code might have
broken due to breaking changes or the tweaks that I have made.

I renamed it to Jäger is because I didn't want the name of the bot to be a meme anymore. Because, memes
move on and so the name will lose its novelty quickly. Why Jäger? Go figure. I'm not telling you.