Fix LeaveCommand to only invoke when bot is connected
The bot was executing the entire command even when the bot was not
connected, due to a lack of proper logic in the command. This was
leading to some amount of user confusion. The bot will now display an
error message if the user attempts to disconnect the bot where there was
no connection anyway. Also, the handling of the disconnection is now
more graceful.